Reputation Suggestion + Language and Clarification
Ive added a suggestion regarding a Reputation System. Id like it if several people were able to review it and provide input on what they think about such a system being used to handle the penalty and reward aspect of the system. I also made many edits to help with language and clarification.