Major reorganization
We are removing the non-technical papers and moving it to the wiki. They
are in non-technical folder for now.
Readme will be written by Kyle.
In the technical sense there will be only two papers: overview and
protocol.
Overview is to settle the strong ideas.
Protocol is where the official paper is going to be.